+++ sum_dice_par/sum_dice.go
@@ -21,7 +21,7 @@ func CalculateProbabilities(params dice_roller.DiceRollParameters) (*dice_roller
 
 func collectAllRolls(value int, roller *dice_roller.DiceRoller) []dice_roller.DiceRoll {
 	rootRoll := initDiceRoll(value, roller)
-	index := make(map[int]dice_roller.DiceRoll)
+	index := make(map[uint64]dice_roller.DiceRoll)
 	collectAllRollsRecursive(roller, rootRoll, index)
 	var rolls []dice_roller.DiceRoll
 	for _, roll := range index {
@@ -30,7 +30,9 @@ func collectAllRolls(value int, roller *dice_roller.DiceRoller) []dice_roller.Di
 	return rolls
 }
 
-func collectAllRollsRecursive(roller *dice_roller.DiceRoller, roll dice_roller.DiceRoll, index map[int]dice_roller.DiceRoll) {
+func collectAllRollsRecursive(
+	roller *dice_roller.DiceRoller, roll dice_roller.DiceRoll, index map[uint64]dice_roller.DiceRoll,
+) {
 	index[roller.GetRollIdx(roll)] = roll
 	for _, r := range getAllSimilarRolls(roller, roll) {
 		if _, has := index[roller.GetRollIdx(r)]; !has {
